Fixes:    Compiler warnings (the truncation is deliberate)
[git-p4: depot-paths = "//depot/qt/main/": change = 214093]

+++ src/gui/text/qfontengine.cpp
@@ -215,10 +215,10 @@ void QFontEngine::addBitmapFontToPath(qreal x, qreal y, const QGlyphLayout *glyp
 QImage QFontEngine::alphaMapForGlyph(glyph_t glyph)
 {
     glyph_metrics_t gm = boundingBox(glyph);
-    int glyph_x = floor(gm.x.toReal());
-    int glyph_y = floor(gm.y.toReal());
-    int glyph_width = ceil((gm.x + gm.width).toReal()) -  glyph_x + 2;
-    int glyph_height = ceil((gm.y + gm.height).toReal()) - glyph_y + 2;
+    int glyph_x = int(floor(gm.x.toReal()));
+    int glyph_y = int(floor(gm.y.toReal()));
+    int glyph_width = int(ceil((gm.x + gm.width).toReal())) -  glyph_x + 2;
+    int glyph_height = int(ceil((gm.y + gm.height).toReal())) - glyph_y + 2;
 
     if (glyph_width <= 0 || glyph_height <= 0)
         return QImage();
